Add exception form component
